结构调整

+++ src/main/java/org/ccframe/subsys/core/dto/OrgRowDto.java
@@ -10,7 +10,7 @@ public class OrgRowDto extends Org {
 	private static final long serialVersionUID = 565041499666817293L;
 
 	private Double chargeTotalValue; // 充值总金额
-	private Integer cyclingNum; // 骑行订单数
+	private Long cyclingNum; // 骑行订单数
 	private Double cyclingIncome; // 骑行总收入
 
 	public Double getChargeTotalValue() {
@@ -21,11 +21,11 @@ public class OrgRowDto extends Org {
 		this.chargeTotalValue = chargeTotalValue;
 	}
 
-	public Integer getCyclingNum() {
+	public Long getCyclingNum() {
 		return cyclingNum;
 	}
 
-	public void setCyclingNum(Integer cyclingNum) {
+	public void setCyclingNum(Long cyclingNum) {
 		this.cyclingNum = cyclingNum;
 	}
